+++ src/main/java/sa/gov/nic/bio/bw/client/core/utils/AppConstants.java
@@ -1,5 +1,7 @@
 package sa.gov.nic.bio.bw.client.core.utils;
 
+import sa.gov.nic.bio.bw.client.AppEntryPoint;
+
 import java.time.ZoneId;
 import java.util.Locale;
 
@@ -11,7 +13,8 @@ public interface AppConstants
 		Locale SAUDI_EN_LOCALE = new Locale("en", "SA");
 	}
 	
-	String LOGGER_NAME = "sa.gov.nic.bio.bw.client";
+	Class<?> PREF_NODE_CLASS = AppEntryPoint.class;
+	String UI_LANGUAGE_PREF_NAME = "sa.gov.nic.bio.bw.ui.language";
 	ZoneId SAUDI_ZONE = ZoneId.of("GMT+3");
 	String LOGS_FOLDER_PATH = "C:/bio/apps/bw/logs";
 	String TEMP_FOLDER_PATH = "C:/bio/apps/bw/temp";