+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@
 - Show the environment type in the window title.
 - Add "change-password" feature.
 - Make the change-password-thread daemon.
+- Start the application with the OS's language in the first run and save the language in user preferences on switching for subsequent launches of the application.
 
 ## v1.1.2